+++ app/src/main/java/com/example/vinyls/artist/ArtistFragment.kt
@@ -1,4 +1,4 @@
-package com.example.vinyls.ui.home
+package com.example.vinyls.artist
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,11 +7,11 @@ import android.view.ViewGroup
 import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
-import com.example.vinyls.databinding.FragmentHomeBinding
+import com.example.vinyls.databinding.FragmentArtistBinding
 
-class HomeFragment : Fragment() {
+class ArtistFragment : Fragment() {
 
-    private var _binding: FragmentHomeBinding? = null
+    private var _binding: FragmentArtistBinding? = null
 
     // This property is only valid between onCreateView and
     // onDestroyView.
@@ -22,14 +22,14 @@ class HomeFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        val homeViewModel =
-            ViewModelProvider(this).get(HomeViewModel::class.java)
+        val artistViewModel =
+            ViewModelProvider(this).get(ArtistViewModel::class.java)
 
-        _binding = FragmentHomeBinding.inflate(inflater, container, false)
+        _binding = FragmentArtistBinding.inflate(inflater, container, false)
         val root: View = binding.root
 
-        val textView: TextView = binding.textHome
-        homeViewModel.text.observe(viewLifecycleOwner) {
+        val textView: TextView = binding.textArtist
+        artistViewModel.text.observe(viewLifecycleOwner) {
             textView.text = it
         }
         return root
