+++ app/src/main/java/com/example/vinyls/home/HomeViewModel.kt
@@ -1,13 +1,13 @@
-package com.example.vinyls.ui.gallery
+package com.example.vinyls.home
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class GalleryViewModel : ViewModel() {
+class HomeViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is gallery Fragment"
+        value = "This is home Fragment"
     }
     val text: LiveData<String> = _text
 }
