+++ app/src/main/res/layout/fragment_artist.xml
@@ -4,10 +4,10 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.slideshow.SlideshowFragment">
+    tools:context=".artist.ArtistFragment">
 
     <TextView
-        android:id="@+id/text_slideshow"
+        android:id="@+id/text_artist"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_marginStart="8dp"
