+++ app/src/main/res/menu/activity_main_drawer.xml
@@ -9,12 +9,12 @@
             android:icon="@drawable/ic_menu_camera"
             android:title="@string/menu_home" />
         <item
-            android:id="@+id/nav_gallery"
+            android:id="@+id/nav_album"
             android:icon="@drawable/ic_menu_gallery"
-            android:title="@string/menu_gallery" />
+            android:title="@string/menu_album" />
         <item
-            android:id="@+id/nav_slideshow"
+            android:id="@+id/nav_artist"
             android:icon="@drawable/ic_menu_slideshow"
-            android:title="@string/menu_slideshow" />
+            android:title="@string/menu_artist" />
     </group>
 </menu>
