+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -7,19 +7,19 @@
 
     <fragment
         android:id="@+id/nav_home"
-        android:name="com.example.vinyls.ui.home.HomeFragment"
+        android:name="com.example.vinyls.home.HomeFragment"
         android:label="@string/menu_home"
         tools:layout="@layout/fragment_home" />
 
     <fragment
-        android:id="@+id/nav_gallery"
-        android:name="com.example.vinyls.ui.gallery.GalleryFragment"
-        android:label="@string/menu_gallery"
-        tools:layout="@layout/fragment_gallery" />
+        android:id="@+id/nav_album"
+        android:name="com.example.vinyls.album.AlbumFragment"
+        android:label="@string/menu_album"
+        tools:layout="@layout/fragment_album" />
 
     <fragment
-        android:id="@+id/nav_slideshow"
-        android:name="com.example.vinyls.ui.slideshow.SlideshowFragment"
-        android:label="@string/menu_slideshow"
-        tools:layout="@layout/fragment_slideshow" />
+        android:id="@+id/nav_artist"
+        android:name="com.example.vinyls.artist.ArtistFragment"
+        android:label="@string/menu_artist"
+        tools:layout="@layout/fragment_artist" />
 </navigation>
