+++ app/src/main/res/values/strings.xml
@@ -8,6 +8,6 @@
     <string name="action_settings">Settings</string>
 
     <string name="menu_home">Home</string>
-    <string name="menu_gallery">Gallery</string>
-    <string name="menu_slideshow">Slideshow</string>
+    <string name="menu_album">Album</string>
+    <string name="menu_artist">Artist</string>
 </resources>
